Export site list parsing to separate file
